1. Added word autocompletion (modified version of internet sample) 2. Edited to fit with existing GUI. (GUI class modified lightly - just the search textfield changed to a global variable) 3. Added line to extract query term from jtextfield (this can be seen printed on the console) - this is done inside GUIController.
Notes:
Word autocompletion is CASE sensitive. 
A list of words that are autocompleted (the "dictionary" is listed
inside GUI) - I've sectioned out the area
Content is NOT displayed on GUI as of yet because search system as has
to be implemented. You can see the query on the console instead.